stdio: Fix UB on snprintf
The elf/tst-dl-printf-static test when built with ubsan triggers:

UBSAN: Undefined behaviour in vfprintf-process-arg.c:58:36 negation of 9223372036854775808 cannot be represented in type 'long int'

+++ stdio-common/vfprintf-process-arg.c
@@ -55,7 +55,8 @@ LABEL (form_integer):
         signed_number = (short int) process_arg_unsigned_int ();
 
       is_negative = signed_number < 0;
-      number.word = is_negative ? (- signed_number) : signed_number;
+      number.word = is_negative ? (- (unsigned long) signed_number)
+				: signed_number;
 
       goto LABEL (number);
     }
